backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2.3
django-cache-url==3.4.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
environs==9.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
marshmallow==3.19.0
msgpack==1.0.5
packaging==23.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.29.1
rpds-py==0.8.10
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
zope.interface==6.0
